ssouk/requirements/project.txt
# project.txt is a pip requirements file which describes the distributions
# required by your project to run.

--requirement=base.txt

# Put project-specific requirements here.
# See http://pip-installer.org/requirement-format.html for more information.


Pillow==1.7.2
South==0.7.3
distribute==0.6.24
django-announcements==0.1.2
django-appconf==0.4.1
#django-floppyforms==0.4.7
 -e git+https://github.com/ssouk/django-notification.git#egg=django-notification
django-pagination==1.0.7
#django-uni-form==0.9.0
gondor==1.0
-e git+http://github.com/eldarion/idios.git#egg=idios-dev
ipython
psycopg2==2.4.1
simplejson==2.1.6
wsgiref==0.1.2
yolk==0.4.1
django-crispy-forms==1.1.3
